refactor: extract shared logic from leaderboard tables
