+++ src/components/leaderboard/TopRepositoriesTable.tsx
@@ -6,7 +6,6 @@ import React, {
   useRef,
 } from 'react';
 import {
-  alpha,
   Box,
   Card,
   Table,
@@ -36,8 +35,13 @@ import BarChartIcon from '@mui/icons-material/BarChart';
 import TableChartIcon from '@mui/icons-material/TableChart';
 import ReactECharts from 'echarts-for-react';
 import { useSearchParams } from 'react-router-dom';
-import { RANK_COLORS } from '../../theme';
 import { truncateText } from '../../utils';
+import {
+  bodyCellStyle,
+  getRankIcon,
+  headerCellStyle,
+} from './tableSharedUtils';
+import { getRepositoryChartOption } from './chartFormatters';
 
 interface RepoStats {
   repository: string;
@@ -179,196 +183,6 @@ const TopRepositoriesTable: React.FC<TopRepositoriesTableProps> = ({
 
     return filtered;
   }, [rankedRepositories, searchQuery]);
-
-  const getChartOption = () => {
-    const chartData = filteredRepositories.slice(0, 50); // Limit for performance
-    const textColor = 'rgba(255, 255, 255, 0.85)';
-    const gridColor = 'rgba(255, 255, 255, 0.08)';
-
-    const barGradient = {
-      type: 'linear',
-      x: 0,
-      y: 0,
-      x2: 0,
-      y2: 1,
-      colorStops: [
-        { offset: 0, color: 'rgba(139, 148, 158, 0.8)' },
-        { offset: 0.5, color: 'rgba(139, 148, 158, 0.6)' },
-        { offset: 1, color: 'rgba(100, 108, 118, 0.4)' },
-      ],
-    };
-
-    const xAxisData = chartData.map((item) => ({
-      name: (item?.repository || '').split('/')[1] || item?.repository || '',
-      fullName: item?.repository || '',
-    }));
-
-    const seriesData = chartData.map((item, index) => ({
-      value: Number(item?.totalScore) || 0,
-      rank: item?.rank || index + 1,
-      repository: item?.repository || '',
-      weight: item?.weight || 0,
-      prs: item?.totalPRs || 0,
-      contributors: item?.uniqueMiners?.size || 0,
-      itemStyle: {
-        color: barGradient,
-        borderRadius: [6, 6, 0, 0],
-        shadowColor: 'rgba(100, 100, 100, 0.2)',
-        shadowBlur: 12,
-      },
-    }));
-
-    return {
-      backgroundColor: 'transparent',
-      title: {
-        text: 'Repository Score Performance',
-        subtext: 'Total score generated by repository contributions',
-        left: 'center',
-        top: 20,
-        textStyle: {
-          color: '#ffffff',
-          fontFamily: 'JetBrains Mono',
-          fontSize: 18,
-          fontWeight: 600,
-        },
-        subtextStyle: {
-          color: 'rgba(255, 255, 255, 0.5)',
-          fontFamily: 'JetBrains Mono',
-          fontSize: 12,
-        },
-      },
-      tooltip: {
-        trigger: 'axis',
-        axisPointer: {
-          type: 'shadow',
-          shadowStyle: {
-            color: 'rgba(255, 255, 255, 0.05)',
-          },
-        },
-        backgroundColor: 'rgba(15, 15, 18, 0.95)',
-        borderColor: 'rgba(255, 255, 255, 0.15)',
-        borderWidth: 1,
-        textStyle: {
-          color: '#fff',
-          fontFamily: 'JetBrains Mono',
-          fontSize: 12,
-        },
-        padding: [12, 16],
-        formatter: (params: any) => {
-          const data = params[0];
-          const item = seriesData[data.dataIndex];
-
-          return `
-            <div style="font-family: 'JetBrains Mono', monospace;">
-              <div style="font-weight: 600; margin-bottom: 8px; font-size: 13px;">
-                #${item.rank} ${item.repository}
-              </div>
-              <div style="margin-top: 8px; padding-top: 8px; border-top: 1px solid rgba(255,255,255,0.1);">
-                <div style="color: rgba(255,255,255,0.7); margin-bottom: 4px;">Total Score: <span style="color: #fff; font-weight: 600;">${item.value.toFixed(2)}</span></div>
-                <div style="color: rgba(255,255,255,0.7); margin-bottom: 4px;">Weight: <span style="color: #fff; font-weight: 600;">${item.weight.toFixed(2)}</span></div>
-                <div style="color: rgba(255,255,255,0.7); margin-bottom: 4px;">Pull Requests: <span style="color: #fff; font-weight: 600;">${item.prs}</span></div>
-                <div style="color: rgba(255,255,255,0.7);">Contributors: <span style="color: #fff; font-weight: 600;">${item.contributors}</span></div>
-              </div>
-            </div>
-          `;
-        },
-      },
-      grid: {
-        left: '3%',
-        right: '3%',
-        bottom: '18%',
-        top: '18%',
-        containLabel: true,
-      },
-      dataZoom: [
-        {
-          type: 'inside',
-          start: 0,
-          end: 100,
-          zoomOnMouseWheel: true,
-          moveOnMouseMove: true,
-        },
-      ],
-      xAxis: {
-        type: 'category',
-        data: xAxisData.map((item) => item.name),
-        axisLabel: {
-          color: textColor,
-          fontFamily: 'JetBrains Mono',
-          fontSize: 11,
-          interval: 0,
-          rotate: 45,
-          margin: 12,
-          formatter: (label: string) =>
-            label.length > 15 ? `${label.substring(0, 12)}...` : label,
-        },
-        axisLine: {
-          lineStyle: {
-            color: gridColor,
-            width: 1,
-          },
-        },
-        axisTick: {
-          show: false,
-        },
-      },
-      yAxis: {
-        type: useLogScale ? 'log' : 'value',
-        min: useLogScale ? 1 : 0,
-        logBase: 10,
-        name: 'Total Score',
-        nameTextStyle: {
-          color: textColor,
-          fontFamily: 'JetBrains Mono',
-          fontSize: 12,
-          padding: [0, 0, 0, 0],
-        },
-        axisLabel: {
-          color: textColor,
-          fontFamily: 'JetBrains Mono',
-          fontSize: 11,
-          formatter: (value: number) => {
-            if (value >= 1000) return `${(value / 1000).toFixed(1)}k`;
-            return value.toFixed(0);
-          },
-        },
-        splitLine: {
-          lineStyle: {
-            color: gridColor,
-            type: 'dashed',
-            opacity: 0.5,
-          },
-        },
-        axisLine: {
-          show: false,
-        },
-        axisTick: {
-          show: false,
-        },
-      },
-      series: [
-        {
-          data: seriesData,
-          type: 'bar',
-          barWidth: '60%',
-          showBackground: true,
-          backgroundStyle: {
-            color: 'rgba(255, 255, 255, 0.02)',
-            borderRadius: [6, 6, 0, 0],
-          },
-          emphasis: {
-            focus: 'series',
-            itemStyle: {
-              shadowBlur: 20,
-              shadowColor: 'rgba(88, 166, 255, 0.5)',
-            },
-          },
-          animationDuration: 1000,
-          animationEasing: 'cubicOut',
-        },
-      ],
-    };
-  };
 
   const handleChangePage = (_event: unknown, newPage: number) => {
     setPage(newPage);
@@ -641,7 +455,10 @@ const TopRepositoriesTable: React.FC<TopRepositoriesTableProps> = ({
         >
           {showChart && filteredRepositories.length > 0 && (
             <ReactECharts
-              option={getChartOption()}
+              option={getRepositoryChartOption(
+                filteredRepositories,
+                useLogScale,
+              )}
               style={{ height: '100%', width: '100%' }}
             />
           )}
@@ -916,83 +733,5 @@ const TopRepositoriesTable: React.FC<TopRepositoriesTableProps> = ({
     </Card>
   );
 };
-
-const headerCellStyle = {
-  backgroundColor: 'rgba(18, 18, 20, 0.95)',
-  backdropFilter: 'blur(8px)',
-  color: '#ffffff',
-  fontFamily: '"JetBrains Mono", monospace',
-  fontWeight: 500,
-  fontSize: '0.75rem',
-  borderBottom: '1px solid rgba(255, 255, 255, 0.1)',
-  height: '48px',
-  py: 1,
-  boxSizing: 'border-box' as const,
-};
-
-const bodyCellStyle = {
-  color: '#ffffff',
-  fontFamily: '"JetBrains Mono", monospace',
-  borderBottom: '1px solid rgba(255, 255, 255, 0.1)',
-  fontSize: '0.75rem',
-  py: 0.75,
-  height: '52px',
-  boxSizing: 'border-box' as const,
-};
-
-const getRankIcon = (rank: number) => (
-  <Box
-    sx={{
-      backgroundColor: '#000000',
-      borderRadius: '2px',
-      width: '22px',
-      height: '22px',
-      display: 'inline-flex',
-      alignItems: 'center',
-      justifyContent: 'center',
-      flexShrink: 0,
-      border: '1px solid',
-      borderColor:
-        rank === 1
-          ? alpha(RANK_COLORS.first, 0.4)
-          : rank === 2
-            ? alpha(RANK_COLORS.second, 0.4)
-            : rank === 3
-              ? alpha(RANK_COLORS.third, 0.4)
-              : 'rgba(255, 255, 255, 0.15)',
-      boxShadow:
-        rank === 1
-          ? `0 0 12px ${alpha(RANK_COLORS.first, 0.4)}, 0 0 4px ${alpha(RANK_COLORS.first, 0.2)}`
-          : rank === 2
-            ? `0 0 12px ${alpha(RANK_COLORS.second, 0.4)}, 0 0 4px ${alpha(RANK_COLORS.second, 0.2)}`
-            : rank === 3
-              ? `0 0 12px ${alpha(RANK_COLORS.third, 0.4)}, 0 0 4px ${alpha(RANK_COLORS.third, 0.2)}`
-              : 'none',
-    }}
-  >
-    <Typography
-      component="span"
-      sx={{
-        color:
-          rank === 1
-            ? RANK_COLORS.first
-            : rank === 2
-              ? RANK_COLORS.second
-              : rank === 3
-                ? RANK_COLORS.third
-                : 'rgba(255, 255, 255, 0.6)',
-        fontFamily: '"JetBrains Mono", monospace',
-        fontSize: '0.65rem',
-        fontWeight: 600,
-        lineHeight: 1,
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'center',
-      }}
-    >
-      {rank}
-    </Typography>
-  </Box>
-);
 
 export default TopRepositoriesTable;
